+++ node/src/utils/external.rs
@@ -22,7 +22,11 @@ use thiserror::Error;
 use casper_types::SecretKey;
 
 use super::{read_file, ReadFileError};
-use crate::{crypto, crypto::AsymmetricKeyExt, tls};
+use crate::{
+    crypto,
+    crypto::AsymmetricKeyExt,
+    tls::{self, LoadCertError, LoadSecretKeyError},
+};
 
 /// Path to bundled resources.
 #[cfg(test)]
@@ -134,7 +138,16 @@ impl Loadable for X509 {
     type Error = anyhow::Error;
 
     fn from_path<P: AsRef<Path>>(path: P) -> Result<Self, Self::Error> {
-        tls::load_cert(path)
+        let error = match tls::load_cert(path) {
+            Ok(cert) => return Ok(cert),
+            Err(LoadCertError::ReadFile(error)) => {
+                anyhow::Error::new(error).context("failed to load certificate")
+            }
+            Err(LoadCertError::X509CertFromPem(error)) => {
+                anyhow::Error::new(error).context("parsing certificate")
+            }
+        };
+        Err(error)
     }
 }
 
@@ -142,7 +155,17 @@ impl Loadable for PKey<Private> {
     type Error = anyhow::Error;
 
     fn from_path<P: AsRef<Path>>(path: P) -> Result<Self, Self::Error> {
-        tls::load_private_key(path)
+        let error = match tls::load_secret_key(path) {
+            Ok(secret_key) => return Ok(secret_key),
+            Err(LoadSecretKeyError::ReadFile(error)) => {
+                anyhow::Error::new(error).context("failed to load private key")
+            }
+            Err(LoadSecretKeyError::PrivateKeyFromPem(error)) => {
+                anyhow::Error::new(error).context("parsing private key")
+            }
+        };
+
+        Err(error)
     }
 }